feat(63-01): extract csv-discovery module and remove block.json from sync layer
- Create shared csv-discovery.ts with detectCsvType, discoverCsvFiles, logCsvDiscoveryWarning
- Refactor block-sync.ts to use discoverCsvFiles instead of readBlockJson
- Remove CsvMappings, BlockJson, readBlockJson from block-sync.ts
- Re-export csv-discovery functions from block-loader.ts for backward compat
- Add csv-discovery exports to test-exports.ts
