+++ packages/mcp-server/src/utils/block-loader.ts
@@ -14,14 +14,9 @@ import {
   convertTatRowToReportingTrade,
 } from "@tradeblocks/lib";
 
-/**
- * CSV file mappings for flexible discovery
- */
-export interface CsvMappings {
-  tradelog?: string;
-  dailylog?: string;
-  reportinglog?: string;
-}
+// Re-export CSV discovery types and functions from shared module
+export { type CsvMappings, type CsvType, detectCsvType, discoverCsvFiles, logCsvDiscoveryWarning } from "./csv-discovery.js";
+import { type CsvMappings, type CsvType, discoverCsvFiles, logCsvDiscoveryWarning } from "./csv-discovery.js";
 
 /**
  * Block metadata stored in block.json
@@ -225,199 +220,7 @@ function parseCSVLine(line: string): string[] {
   return result;
 }
 
-/**
- * CSV type detection result
- */
-type CsvType = "tradelog" | "dailylog" | "reportinglog" | null;
-
-/**
- * Read just the header line from a CSV file (for detection)
- */
-async function readCsvHeaders(filePath: string): Promise<string[]> {
-  const content = await fs.readFile(filePath, "utf-8");
-  const firstLine = content.split("\n")[0] || "";
-  return parseCSVLine(firstLine).map((h) => h.toLowerCase().trim());
-}
-
-/**
- * Detect CSV type by examining column headers.
- * Returns the detected type or null if unrecognized.
- */
-async function detectCsvType(filePath: string): Promise<CsvType> {
-  try {
-    const headers = await readCsvHeaders(filePath);
-
-    // Trade log detection:
-    // Required: "P/L" or "P&L" or "Profit/Loss"
-    // Plus at least 2 of: "Date Opened", "Date Closed", "Symbol", "Strategy", "Contracts", "Premium"
-    const plColumnAliases = ["p/l", "p&l", "profit/loss", "pl"];
-    const tradeOptionalColumns = [
-      "date opened",
-      "date closed",
-      "symbol",
-      "strategy",
-      "contracts",
-      "no. of contracts",
-      "premium",
-      "legs",
-    ];
-
-    const hasPl = plColumnAliases.some((alias) => headers.includes(alias));
-    // Match trade columns - require header to contain the full column pattern
-    // This prevents "date" from matching "date opened" (col.includes(h) would be true)
-    const matchedTradeColumns = tradeOptionalColumns.filter((col) =>
-      headers.some((h) => h.includes(col))
-    );
-
-    if (hasPl && matchedTradeColumns.length >= 2) {
-      return "tradelog";
-    }
-
-    // Daily log detection:
-    // Required: "Date" (but not "Date Opened"/"Date Closed"), and value column
-    // Key distinction: dailylogs have portfolio value columns but lack trade-specific columns
-    const hasSimpleDate = headers.some(
-      (h) => h === "date" || (h.includes("date") && !h.includes("opened") && !h.includes("closed"))
-    );
-    const valueColumnAliases = [
-      "portfolio value",
-      "value",
-      "equity",
-      "net liquidity",
-      "netliquidity",
-    ];
-    const hasValue = valueColumnAliases.some((alias) =>
-      headers.some((h) => h.includes(alias) || alias.includes(h))
-    );
-
-    // Dailylog: has date + value columns but lacks trade-specific columns
-    // This catches dailylogs that also have P/L columns (like Option Omega exports)
-    if (hasSimpleDate && hasValue && matchedTradeColumns.length < 2) {
-      return "dailylog";
-    }
-
-    // TAT (Trade Automation Toolbox) detection:
-    // Has "TradeID" AND "ProfitLoss" AND "BuyingPower"
-    const tatSignature = ['tradeid', 'profitloss', 'buyingpower'];
-    const isTat = tatSignature.every((sig) => headers.includes(sig));
-    if (isTat) {
-      return 'reportinglog';
-    }
-
-    // Reporting log detection:
-    // Has "Actual P/L" or columns from REPORTING_TRADE_COLUMN_ALIASES
-    // Or has "Trade ID" + "Reported" style columns
-    const reportingAliases = Object.keys(REPORTING_TRADE_COLUMN_ALIASES).map(
-      (k) => k.toLowerCase()
-    );
-    const hasReportingColumns = reportingAliases.some((alias) =>
-      headers.includes(alias)
-    );
-    const hasActualPl = headers.some(
-      (h) => h.includes("actual") && h.includes("p")
-    );
-    const hasReportedStyle =
-      headers.includes("trade id") ||
-      headers.some((h) => h.includes("reported"));
-
-    if (hasActualPl || hasReportingColumns || hasReportedStyle) {
-      // Double-check it's not a regular tradelog
-      if (!hasPl || hasActualPl) {
-        return "reportinglog";
-      }
-    }
-
-    // If we have P/L and trade columns, fallback to tradelog
-    if (hasPl && matchedTradeColumns.length >= 1) {
-      return "tradelog";
-    }
-
-    return null;
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Discover CSV files in a folder and detect their types.
- * Returns mapping of detected CSV types to filenames.
- */
-async function discoverCsvFiles(
-  folderPath: string
-): Promise<{ mappings: CsvMappings; unrecognized: string[] }> {
-  const mappings: CsvMappings = {};
-  const unrecognized: string[] = [];
-
-  try {
-    const entries = await fs.readdir(folderPath, { withFileTypes: true });
-    const csvFiles = entries
-      .filter((e) => e.isFile() && e.name.toLowerCase().endsWith(".csv"))
-      .map((e) => e.name);
-
-    // First, check for exact standard names
-    if (csvFiles.includes("tradelog.csv")) {
-      mappings.tradelog = "tradelog.csv";
-    }
-    if (csvFiles.includes("dailylog.csv")) {
-      mappings.dailylog = "dailylog.csv";
-    }
-    if (csvFiles.includes("reportinglog.csv")) {
-      mappings.reportinglog = "reportinglog.csv";
-    }
-
-    // Second, check for filename patterns (before content detection)
-    // This helps with Option Omega exports where strategy-trade-log has same columns as tradelog
-    if (!mappings.reportinglog) {
-      const strategyLogFile = csvFiles.find((f) => {
-        const lower = f.toLowerCase();
-        return (
-          lower.includes("strategy-trade-log") ||
-          lower.includes("strategylog") ||
-          lower.startsWith("strategy-log")
-        );
-      });
-      if (strategyLogFile) {
-        mappings.reportinglog = strategyLogFile;
-      }
-    }
-
-    // For any remaining CSVs, detect by content
-    for (const csvFile of csvFiles) {
-      // Skip if already mapped via exact name or filename pattern
-      if (
-        csvFile === "tradelog.csv" ||
-        csvFile === "dailylog.csv" ||
-        csvFile === "reportinglog.csv" ||
-        csvFile === mappings.reportinglog
-      ) {
-        continue;
-      }
-
-      const csvPath = path.join(folderPath, csvFile);
-      const detectedType = await detectCsvType(csvPath);
-
-      if (detectedType) {
-        // Only assign if we haven't found this type yet
-        if (detectedType === "tradelog" && !mappings.tradelog) {
-          mappings.tradelog = csvFile;
-        } else if (detectedType === "dailylog" && !mappings.dailylog) {
-          mappings.dailylog = csvFile;
-        } else if (detectedType === "reportinglog" && !mappings.reportinglog) {
-          mappings.reportinglog = csvFile;
-        } else {
-          // Type already found, this is an extra CSV
-          unrecognized.push(csvFile);
-        }
-      } else {
-        unrecognized.push(csvFile);
-      }
-    }
-  } catch {
-    // Folder read error - return empty
-  }
-
-  return { mappings, unrecognized };
-}
+// Note: CsvType, readCsvHeaders, detectCsvType, discoverCsvFiles moved to csv-discovery.ts
 
 /**
  * Get modification times for CSV files in a block folder.
@@ -490,17 +293,7 @@ async function isCacheValid(
   return true;
 }
 
-/**
- * Log warning when folder has CSVs but none match expected patterns
- */
-function logCsvDiscoveryWarning(
-  folderName: string,
-  csvFiles: string[]
-): void {
-  console.error(`Warning: Folder '${folderName}' has CSV files but none match expected trade log format.`);
-  console.error(`  Found: ${csvFiles.join(", ")}`);
-  console.error(`  Expected columns: P/L, Date Opened, Date Closed, Symbol, Strategy`);
-}
+// Note: logCsvDiscoveryWarning moved to csv-discovery.ts
 
 /**
  * Convert raw CSV record to Trade object
